feat: use animation for x loading refactor: move login contents to its own components
