+++ package.json
@@ -98,6 +98,7 @@
     "react-infinite-scroll-component": "^6.1.0",
     "react-intersection-observer": "^9.4.3",
     "react-lite-youtube-embed": "^2.3.52",
+    "react-lottie": "^1.2.3",
     "react-qr-code": "^2.0.11",
     "react-social-media-embed": "^2.3.5",
     "react-tweet": "^3.1.0",
@@ -133,6 +134,7 @@
     "@subsocial/config": "^1.0.0",
     "@svgr/webpack": "^6.5.1",
     "@total-typescript/ts-reset": "^0.5.1",
+    "@types/react-lottie": "^1.2.10",
     "autoprefixer": "^10.4.13",
     "dotenv": "^16.0.3",
     "eslint-plugin-storybook": "^0.6.13",
